feestimator: encapsulate estimation file logic
This makes the handling of the fee estimation file atomic with the
CBlockPolicyEstimator component.

The path to the estimation file is now passed to its constructor in
which it handles reading it if it exists (and now also log if not).

It will handle the still-unconfirmed txs flushing and
current-estimations writing in its destructor.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>