qa: add a fuzz target generating random nodes from a binary encoding
This is a "dumb" version of the Miniscript node generation from fuzzer
input. It defines a strict binary encoding and will always generate a
node defined from the encoding without "helping" to create valid nodes.
It will cut through as soon as it encounters an invalid fragment so
hopefully the fuzzer can tend to learn the encoding and generate valid
nodes with a higher probability.

On a valid generated node a number of invariants are checked, especially
around the satisfactions and testing them against the Script
interpreter.

Co-Authored-By: Pieter Wuille <pieter@wuille.net>