miniscript: allow to delay the duplicate key check in the ctor
This makes the Node constructors private, and makes them not perform the
check for duplicate keys. Instead, 'duplicate_key' is made mutable and
the check is done only once on the top level node after parsing.

We provide a public static method 'Node::New()' to create a Node which
always performs the check for duplicate keys.

Unfortunately, we can't use make_shared with private ctors. This makes
BuildBack and MakeNodeRef friend functions that instead use 'new' to
allocate the Nodes when parsing.