consesus/mekle: don't perform the ignored malleation check for witnesses
The result of the malleation check is ignored, since it's not possible
to exploit it for the witness tree without exploiting it for the
transaction tree (where we do check it).

The check involves going through all the leaves and nodes of the tree
and comparing them. Since the result is ignored, we better not perform
the check at all.

Since the ignored `malleated` boolean was the only call site, remove the
argument altogether.