qa: make feature_assumeutxo.py test more robust
The feature_assumeutxo.py functional test contains a set of checks for expected errors when a
snapshot can be parsed but contains incorrect values. Some of these checks are brittle as they
assert an error that is raised not only because of the injected fault, but also because of the
following data in the snapshot. A previous occurence of this was fixed in
e5ff4e416ecc8a51367022eb8a7a291f8cbc0c65.

In this commit, we look into another check of the same category. This one inserts an invalid coins
count for the first transaction in the snapshot. It asserts that by doing so the snapshot validation
logic will try to deserialize an additional coin for the first transaction, while it fact it will be
deserializing the txid of the next transaction. It then asserts that this attempt to deserialize the
txid of the next transaction will result in a specific error: a coin height higher than the base
height (returning the string "Bad snapshot data after deserializing 1 coins.").

This is an issue as it relies on data outside of the fault injected to assert the error string. For
instance the next txid could well deserialize in a coin at a height lower than the base height, or
(more probably) not successfully deserialize into a coin at all!

Fix this by injecting all the data necessary to reliably assert the expected error: an invalid coins
count, followed by a valid coin, followed by a valid-serialization coin with an invalid height
(which in this scenario is supposed to be the next coin's txid).