miner: set the coinbase nLockTime to block height-1
The Consensus Cleanup soft fork proposal includes enforcing that coinbase transactions set their
locktime field to the block height, minus 1 (as well as their nSequence such as not to disable the
timelock). If such a fork were to be activated by Bitcoin users, miners need to be ready to produce
compliant blocks at the risk of losing substantial amounts mining would-be invalid blocks.

Bitcoin Core is the most widely used block template provider. Miners are known to be slow to upgrade
to newer versions of the software. Setting the coinbase's nLockTime to the block height minus 1
instead of 0 is therefore a cost-free way of hedging against an activation of the Consensus Cleanup
in the future.

A nice side-effect is that it will allow Bitcoin users to track when it's safe to attempt activating
such a soft fork.