miniscript: introduce a pk_i() fragment for OP_INTERNALKEY
This is a bit convoluted since no other fragment relied on the Taproot
internal key yet, and it needs to be passed from the context, and
therefore all places where we create a context: in signing, descriptor
parsing / inference, unit tests and fuzz tests.

The approach taken is to only query the internal key once at parsing /
inference time and store it in the fragment's keys vector. This way it
naturally integrates with the existing code for pk_k(), such for
serialization, signing and duplicate key checks.

Care was taken in fuzz harnesses to not invalidate existing seeds, but
also use a meaningful key in TestNode() (i.e. in miniscript_smart and
miniscript_stable targets). A new optional internal_key parameter is
adding down the call chain as an "out" parameter: the first time the
fuzzer generates a pk_i() fragment (if any) the internal key is set and
reused for all pk_i() fragments. This way we can roundtrip ser/parsing,
make it available to the satisfier to sign, etc..