miniscript: introduce a 'th(hash)' fragment
This fragment checks the spending transaction's template hash. Because
it commits to the spending transaction (with more malleable fields than
SIGHASH_ALL, but less than SIGHASH_SINGLE/NONE) then we give it the 't'
property. Note this breaks the invariant that a 't' fragment always
contains at least one key.

In the miniscript_smart and miniscript_stable fuzz harnesses we use the
message hash for dummy signatures as the template hash. We sometimes
create th() fragments with this hash (making them satisfiable), and
sometimes not.